requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==4.2.23
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
django-taggit==6.1.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
